+++ src/lib/kotlin/slatekit-apis/src/main/kotlin/slatekit/apis/core/Errs.kt
@@ -1,4 +1,4 @@
-package slatekit.apis.services
+package slatekit.apis.core
 
 import org.json.simple.JSONArray
 import org.json.simple.JSONObject
