+++ src/lib/kotlin/slatekit-apis/src/main/kotlin/slatekit/apis/services/Validate.kt
@@ -1,8 +1,7 @@
-package slatekit.apis.hooks
+package slatekit.apis.services
 
 import slatekit.apis.ApiRequest
 import slatekit.apis.Middleware
-import slatekit.apis.services.Calls
 import slatekit.common.Ignore
 import slatekit.results.Outcome
